Use clap 4 (#172)
Clap v4 is faster, and has a good migration guide.

I changed the openapitor CLI slightly. It previously had a --version flag which was used for setting the version of the crate which it output. But in clap v4, this conflicts with the builtin --version flag which shows the version of openapitor itself.

To reduce confusion I renamed the flag --target-version.