Bump tabled crate 0.8 -> 0.12
This changes some of the Tabled trait type signatures to return
copy-on-write strings. They're more efficient because they can be a
reference to a static string. In many places in this PR I'm just using
cloned owned strings instead of references to static strings, but this
will do for now. In the future we can improve it.